feat: add jack parser
