--- src/abstractions/software/compiler/mockData/SyntaxAnalayzer/index.js
@@ -1,23 +1,0 @@
-import {
-  ARRAY_TEST_JACK_CODE,
-  ARRAY_TEST_TOKENS
-} from './ArrayTest'
-import {
-  SQUARE_MAIN,
-  SQUARE_MAIN_TOKENS,
-  SQUARE_SQUARE,
-  SQUARE_SQUARE_TOKENS,
-  SQUARE_SQUARE_GAME,
-  SQUARE_SQUARE_GAME_TOKENS
-} from './Square'
-
-export default {
-  ARRAY_TEST_JACK_CODE,
-  ARRAY_TEST_TOKENS,
-  SQUARE_MAIN,
-  SQUARE_MAIN_TOKENS,
-  SQUARE_SQUARE,
-  SQUARE_SQUARE_TOKENS,
-  SQUARE_SQUARE_GAME,
-  SQUARE_SQUARE_GAME_TOKENS
-}
