+++ src/abstractions/software/compiler/mockData/index.js
@@ -1,4 +1,5 @@
-import SyntaxAnalayzerMockData from './SyntaxAnalayzer'
-import ExpressionLessMockData from './parser'
+import * as ArrayTest from './ArrayTest'
+import * as ExpressionLessSquare from './ExpressionLessSquare'
+import * as Square from './Square'
 
-export { SyntaxAnalayzerMockData, ExpressionLessMockData }
+export { ArrayTest, ExpressionLessSquare, Square }
