--- src/abstractions/software/compiler/mockData/parser/index.js
@@ -1,17 +1,0 @@
-import {
-  SQUARE_MAIN_JACK,
-  SQUARE_MAIN_PARSE,
-  SQUARE_SQUARE_GAME_JACK,
-  SQUARE_SQUARE_GAME_PARSE,
-  SQUARE_SQUARE_JACK,
-  SQUARE_SQUARE_PARSE
-} from './expressionLess/square'
-
-export default {
-  SQUARE_MAIN_JACK,
-  SQUARE_MAIN_PARSE,
-  SQUARE_SQUARE_GAME_JACK,
-  SQUARE_SQUARE_GAME_PARSE,
-  SQUARE_SQUARE_JACK,
-  SQUARE_SQUARE_PARSE
-}
